backend/python_rag/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
rank-bm25
pymupdf
pandas
openpyxl
python-dotenv
pydantic
pydantic-settings
numpy
scikit-learn
tiktoken
